% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\docType{data}
\name{spectra}
\alias{spectra}
\title{Spectroscopic scan data}
\format{
A data frame with columns:
\describe{
  \item{freq}{Scan frequencies in Hz.}
  \item{readout}{Fluorescence readout values.}
}
}
\source{
Experimental data from a Paul trap spectroscopy setup.
Collected on 2025-05-23.
}
\usage{
data(spectra)
}
\description{
A single spectroscopic scan from a Paul trap experiment.
}
\examples{
data(spectra)
plot(spectra$freq, spectra$readout)
}
\keyword{datasets}
